facial-recognition-voting/requirements.txt
flask==3.1.0
flask-cors==5.0.0
numpy==2.2.1
opencv-python-headless==4.11.0.86
scikit-learn==0.24.2
pymongo==4.11
python-dotenv==0.19.0
streamlit==1.41.0
pillow==9.0.0
dnspython==2.3.0
pytest==7.4.0
pytest-cov==6.0.0
mongomock==4.3.0